refactor: resolve governance linting rules and strict type definitions
